+++ src/main/scala/services/FormatService.scala
@@ -2,23 +2,13 @@ package services
 
 import model.{FunProgResult, MowerAfterMovement}
 import play.api.libs.json.Json
+import util.YamlConverters.YamlOps
 
 case class FormatService() {
 
   private val csvColumns = {
     "numéro;début_x;début_y;début_direction;fin_x;fin_y;fin_direction;instructions\n"
   }
-  private val instructions = "instructions:"
-  private val directionKey = "direction: "
-  private val basicSpacing = " "
-  private val xKey = "x: "
-  private val yKey = "y: "
-  private val yamlBorder = "---"
-  private val mowerFrench = "tondeuses:"
-  private val garden = "limite:"
-  private val startPoint = "- debut:"
-  private val endPoint = "fin:"
-  private val point = "point:"
 
   def buildCsvOutput(dataToWrite: FunProgResult): String = {
     def getCsvLineFromData(mower: MowerAfterMovement, index: Int): String = {
@@ -41,46 +31,8 @@ case class FormatService() {
     Json.prettyPrint(Json.toJson(dataToWrite))
   }
 
-  def buildYamlOutput(dataToWrite: FunProgResult): List[String] = {
-    val result = List(
-      yamlBorder,
-      garden,
-      basicSpacing + xKey + dataToWrite.limite.x.toString,
-      basicSpacing + yKey + dataToWrite.limite.y.toString,
-      mowerFrench
-    )
-
-    val mowers =
-      dataToWrite.tondeuses.flatMap(mower => mowerToYamlFormat(mower))
-    result.appendedAll(mowers).appended(yamlBorder)
-  }
-
-  private def pointToYamlFormat(
-      x: String,
-      y: String,
-      repeat: Int): List[String] =
-    List(
-      basicSpacing.repeat(repeat).concat(point),
-      basicSpacing.repeat(repeat + 2).concat(xKey + x),
-      basicSpacing.repeat(repeat + 2).concat(yKey + y)
-    )
-
-  private def mowerToYamlFormat(mower: MowerAfterMovement): List[String] = {
-    val (xFin, yFin, direction) = extractFinFromMower(mower)
-
-    List(startPoint) :::
-      pointToYamlFormat(
-        mower.debut.point.x.toString,
-        mower.debut.point.y.toString,
-        4
-      ) :::
-      List(
-        basicSpacing.repeat(4).concat(directionKey + mower.debut.direction.entryName)
-      ) :::
-      instructionsToYaml(mower)
-        .appended(basicSpacing.repeat(2).concat(endPoint)) :::
-      pointToYamlFormat(xFin, yFin, 4)
-        .appended(basicSpacing.repeat(4).concat(directionKey + direction))
+  def buildYamlOutput(dataToWrite: FunProgResult): String = {
+    dataToWrite.toYaml.stringify(0, isFirstLineOfArray = false)
   }
 
   private def extractFinFromMower(mower: MowerAfterMovement): (String, String, String) = {
@@ -91,8 +43,4 @@ case class FormatService() {
         )
   }
 
-  private def instructionsToYaml(mower: MowerAfterMovement): List[String] = {
-    List(basicSpacing.repeat(2).concat(instructions)) ::: mower.instructions
-      .map(instr => "  - " + instr.entryName)
-  }
 }
